+++ bigmler/prediction.py
@@ -371,7 +371,7 @@ def local_batch_predict(models, test_reader, prediction_file, api,
                     model = u.check_resource(model, api.get_model,
                                              query_string)
                 except ValueError, exception:
-                    sys.exit("Failed to get model: %s" % (model,
+                    sys.exit("Failed to get model: %s. %s" % (model,
                                                           str(exception)))
             # When user selects the labels in multi-label predictions, we must
             # filter the models that will be used to predict