+++ bigmler/utils.py
@@ -39,10 +39,11 @@ from bigml.util import console_log
 PAGE_LENGTH = 200
 ATTRIBUTE_NAMES = ['name', 'label', 'description']
 NEW_DIRS_LOG = ".bigmler_dirs"
+BRIEF_MODEL_QS = "exclude=root,fields"
 
 # Base Domain
 BIGML_DOMAIN = os.environ.get('BIGML_DOMAIN', 'bigml.io')
-BIGML_DASHBOARD_URL = os.environ.get('BIGML_DASHBOARD_URL')  
+BIGML_DASHBOARD_URL = os.environ.get('BIGML_DASHBOARD_URL')
 RESOURCE_URL = ("https://%s/dashboard/" % (BIGML_DOMAIN[:-3] + '.com')
                 if BIGML_DASHBOARD_URL is None
                 else BIGML_DASHBOARD_URL)
@@ -109,9 +110,9 @@ def read_types(path):
 
 
 def read_fields_map(path):
-    """Fields map from test dataset to evaluated model.
+    """Fields map from evaluated model to test dataset.
 
-    The test dataset field column and the evaluated model field column
+    The the evaluated model field column and the test dataset field column
     separated by a comma per line.
 
     For example:
@@ -417,3 +418,69 @@ def check_resource(*args, **kwargs):
         return result
     except ValueError, exc:
         sys.exit("\nFailed to obtain a finished resource:\n%s" % str(exc))
+
+
+def objective_field_names(models_or_ensembles, api):
+    """Gets the objective field names for a list of models or ensembles
+
+    """
+    objective_fields = []
+    for model_or_ensemble in models_or_ensembles:
+        name = None
+        if isinstance(model_or_ensemble, dict):
+            status_code = bigml.api.get_status(model_or_ensemble)['code']
+            if status_code == bigml.api.FINISHED:
+                name = objective_field_name(model_or_ensemble, api)
+        if name is None:
+            try:
+                model_id = bigml.api.get_model_id(model_or_ensemble)
+                if model_id is not None:
+                    try:
+                        model = check_resource(model_id, api.get_model,
+                                               query_string=BRIEF_MODEL_QS)
+                    except ValueError, exception:
+                        sys.exit("Failed to get a finished model: %s" %
+                                 str(exception))
+                    name = objective_field_name(model, api)
+            except ValueError:
+                try:
+                    ensemble_id = bigml.api.get_ensemble_id(model_or_ensemble)
+                    if ensemble_id is not None:
+                        try:
+                            ensemble = check_resource(
+                                ensemble_id, api.get_ensemble)
+                        except ValueError, exception:
+                            sys.exit("Failed to get a finished ensemble: %s" %
+                                     str(exception))
+                        name = objective_field_name(ensemble, api)
+                except ValueError:
+                    sys.exit("No valid model or ensemble id")
+        if name is not None:
+            objective_fields.append(name)
+
+    return objective_fields
+
+
+def objective_field_name(model_or_ensemble, api):
+    """Gets the objective field name from a dict model or ensemble structure
+
+    """
+    if not isinstance(model_or_ensemble, dict):
+        return None
+    if not 'object' in model_or_ensemble:
+        return None
+    objective_field = None
+    resource_info = model_or_ensemble['object']
+    if 'objective_field' in resource_info:
+        objective_field = resource_info['objective_field']
+    elif 'objective_fields' in resource_info:
+        objective_field = resource_info['objective_fields'][0]
+    if objective_field is None:
+        return None
+    try:
+        bigml.api.get_ensemble_id(model_or_ensemble)
+        model = model_or_ensemble['object']['models'][0]
+        model = api.get_model(model, query_string=BRIEF_MODEL_QS)
+    except ValueError:
+        model = model_or_ensemble
+    return model['object']['model']['model_fields'][objective_field]['name']