+++ tests/features/basic_test_prediction_steps.py
@@ -9,6 +9,7 @@ from bigmler.bigmler import MONTECARLO_FACTOR
 from bigmler.checkpoint import file_number_of_lines
 from ml_test_prediction_steps import i_create_all_ml_resources
 from ml_test_prediction_steps import i_create_all_ml_resources_and_ensembles
+from ml_test_evaluation_steps import i_create_all_ml_resources_for_evaluation
 
 
 @step(r'I create BigML resources uploading train "(.*?)" file to test "(.*?)" and log predictions in "(.*?)" with "(.*?)" as test field separator$')
@@ -285,19 +286,22 @@ def i_check_create_source(step):
         assert False, str(exc)
 
 
-@step(r'I check that the dataset has been created')
-def i_check_create_dataset(step):
-    dataset_file = "%s%sdataset" % (world.directory, os.sep)
+@step(r'I check that the (test\s|train\s)?dataset has been created')
+def i_check_create_dataset(step, suffix=None):
+    import traceback
+    suffix = "" if suffix is None else "_%s" % suffix[:-1]
+    dataset_file = "%s%sdataset%s" % (world.directory, os.sep, suffix)
     try:
         dataset_file = open(dataset_file, "r")
-        dataset = check_resource(dataset_file.readline().strip(),
+        dataset_id = dataset_file.readline().strip()
+        dataset = check_resource(dataset_id,
                                  world.api.get_dataset)
         world.datasets.append(dataset['resource'])
         world.dataset = dataset
         dataset_file.close()
         assert True
     except Exception, exc:
-        assert False, str(exc)
+        assert False, traceback.format_exc()
 
 
 @step(r'I check that the model has been created')
@@ -368,7 +372,7 @@ def i_check_create_models_in_ensembles(step, in_ensemble=True):
 
 @step(r'I check that the evaluation has been created')
 def i_check_create_evaluation(step):
-    evaluation_file = "%s%sevaluation" % (world.directory, os.sep)
+    evaluation_file = "%s%sevaluations" % (world.directory, os.sep)
     try:
         evaluation_file = open(evaluation_file, "r")
         evaluation = check_resource(evaluation_file.readline().strip(),
@@ -381,8 +385,10 @@ def i_check_create_evaluation(step):
         assert False
 
 
-@step(r'I check that the evaluations have been created')
-def i_check_create_evaluations(step):
+@step(r'I check that the (\d+ )?evaluations have been created')
+def i_check_create_evaluations(step, number_of_evaluations=None):
+    if number_of_evaluations is not None:
+        world.number_of_evaluations = int(number_of_evaluations)
     evaluations_file = "%s%sevaluations" % (world.directory, os.sep)
     evaluation_ids = []
     number_of_lines = 0
@@ -517,7 +523,7 @@ def i_create_all_resources_to_evaluate(step, data=None, output=None):
 
 
 @step(r'I create a BigML source from file "(.*)" with locale "(.*)" storing results in "(.*)"')
-def i_create_all_resources_to_evaluate(step, data=None, locale=None, output=None):
+def i_create_all_resources_to_evaluate_with_locale(step, data=None, locale=None, output=None):
     if data is None or locale is None or output is None:
         assert False
     try:
@@ -541,7 +547,8 @@ def i_have_previous_scenario_or_reproduce_it(step, scenario, kwargs):
                  'scenario_e1': [(i_create_all_resources_to_evaluate, True), (i_check_create_source, False), (i_check_create_dataset, False), (i_check_create_model, False), (i_check_create_evaluation, False)],
                  'scenario_ml_1': [(i_create_all_ml_resources, True), (i_check_create_source, False), (i_check_create_dataset, False), (i_check_create_models, False)],
                  'scenario_ml_6': [(i_create_all_ml_resources, True), (i_check_create_source, False), (i_check_create_dataset, False), (i_check_create_models, False)],
-                 'scenario_mle_1': [(i_create_all_ml_resources_and_ensembles, True), (i_check_create_source, False), (i_check_create_dataset, False), (i_check_create_models_in_ensembles, False)]}
+                 'scenario_mle_1': [(i_create_all_ml_resources_and_ensembles, True), (i_check_create_source, False), (i_check_create_dataset, False), (i_check_create_models_in_ensembles, False)],
+                 'scenario_ml_e1': [(i_create_all_ml_resources_for_evaluation, True), (i_check_create_source, False), (i_check_create_dataset, False), (i_check_create_models, False)]}
     if os.path.exists("./%s/" % scenario):
         assert True
     else: