+++ tests/features/evaluation_test_steps.py
@@ -43,7 +43,27 @@ def given_i_create_bigml_resources_using_dataset_to_evaluate(step, output=None):
         assert False
 
 
-@step(r'I create BigML resources using test file "(.*)" to evaluate a model and log evaluation in "(.*)"')
+@step(r'I create BigML resources using test file "([^"]*)" and a fields map "([^"]*)" to evaluate a model and log evaluation in "(.*)"')
+def i_create_all_resources_to_evaluate_with_model(step, data=None, fields_map=None, output=None):
+    if data is None or fields_map is None or output is None:
+        assert False
+    try:
+        retcode = check_call("bigmler --evaluate --test " + data + " --model " +
+                       world.model['resource'] + " --output " + output +
+                       " --fields-map " + fields_map,
+                       shell=True)
+        if retcode < 0:
+            assert False
+        else:
+            world.directory = os.path.dirname(output)
+            world.folders.append(world.directory)
+            world.output = output
+            assert True
+    except OSError as e:
+        assert False
+
+
+@step(r'I create BigML resources using test file "([^"]*)" to evaluate a model and log evaluation in "(.*)"')
 def i_create_all_resources_to_evaluate_with_model(step, data=None, output=None):
     if data is None or output is None:
         assert False
@@ -143,5 +163,3 @@ def i_check_evaluation_key(step, key=None, value=None):
         evaluation_file_json.close()
     except:
         assert False
-
-